+++ tests/TestCase/Command/LinkScannerCommandTest.php
@@ -181,7 +181,7 @@ class LinkScannerCommandTest extends TestCase
         $expectedConfig['fullBaseUrl'] = $this->fullBaseUrl;
         $this->assertEquals($expectedConfig, $this->LinkScanner->getConfig());
 
-        $differentLines = function ($line) {
+        $differentLines = function (string $line): bool {
             $pattern = sprintf('/^Checking https?:\/\/%s/', preg_quote(get_hostname_from_url($this->fullBaseUrl)));
 
             return substr($line, 0, strlen('Checking')) === 'Checking' && !preg_match($pattern, $line);